複数App対応

+++ ebinaAPI/App/edit.ts
@@ -1,21 +1,23 @@
 import express from "express"
 import fs from 'fs'
-import { authToken } from "../utils/auth"
+import { PROJECT_PATH } from "."
+import { authToken } from "../../utils/auth"
+import { mkdirIfNotExist } from "../../utils/utils"
 
 const editRouter = express.Router()
 
-const BASE_PATH_FILE = './project/js/'
-
-const initFolder = () => {
-  if (!fs.existsSync(BASE_PATH_FILE)) fs.mkdirSync(BASE_PATH_FILE, { recursive: true })
+const initFolder = (appName: string) => {
+  mkdirIfNotExist(`${PROJECT_PATH}/${appName}/js`)
 }
 
 editRouter.post('/js/:path', authToken, (req, res) => {
+  const appName = res.locals.appName
+  if (!appName) return res.sendStatus(400)
   try {
     const { path } = req.params
     if (!path) return res.sendStatus(400)
-    initFolder()
-    const fullPath = BASE_PATH_FILE + path.replaceAll('<', '/')
+    initFolder(appName)
+    const fullPath = `${PROJECT_PATH}/${appName}/js/${path}`
     if (fs.existsSync(fullPath)) return res.sendStatus(409)
     const body = req.body
     const isString = typeof body === 'string'
@@ -33,9 +35,11 @@ editRouter.post('/js/:path', authToken, (req, res) => {
 })
 
 editRouter.get('/js', authToken, (req, res) => {
+  const appName = res.locals.appName
+  if (!appName) return res.sendStatus(400)
   try {
-    initFolder()
-    const dir = fs.readdirSync(BASE_PATH_FILE)
+    initFolder(appName)
+    const dir = fs.readdirSync(`${PROJECT_PATH}/${appName}/js`)
     res.status(200).send(dir)
   } catch (err) {
     console.log(err)
@@ -44,11 +48,13 @@ editRouter.get('/js', authToken, (req, res) => {
 })
 
 editRouter.get('/js/:path', authToken, (req, res) => {
+  const appName = res.locals.appName
+  if (!appName) return res.sendStatus(400)
   try {
     const { path } = req.params
     if (!path) return res.sendStatus(400)
-    initFolder()
-    const fullPath = BASE_PATH_FILE + path.replaceAll('<', '/')
+    initFolder(appName)
+    const fullPath = `${PROJECT_PATH}/${appName}/js/${path}`
     if (!fs.existsSync(fullPath)) return res.sendStatus(404)
     if (fs.statSync(fullPath).isDirectory()) {
       const dir = fs.readdirSync(fullPath)
@@ -64,14 +70,16 @@ editRouter.get('/js/:path', authToken, (req, res) => {
 })
 
 editRouter.patch('/js/:path', authToken, (req, res) => {
+  const appName = res.locals.appName
+  if (!appName) return res.sendStatus(400)
   try {
     const { path } = req.params
     if (!path) return res.sendStatus(400)
     const body = req.body
     const isString = typeof body === 'string'
     if (!isString) return res.sendStatus(400)
-    initFolder()
-    const fullPath = BASE_PATH_FILE + path.replaceAll('<', '/')
+    initFolder(appName)
+    const fullPath = `${PROJECT_PATH}/${appName}/js/${path}`
     if (!fs.existsSync(fullPath)) return res.sendStatus(404)
     fs.writeFile(fullPath, body, {}, (err) => {
       if (err) {
@@ -87,11 +95,13 @@ editRouter.patch('/js/:path', authToken, (req, res) => {
 })
 
 editRouter.delete('/js/:path', authToken, (req, res) => {
+  const appName = res.locals.appName
+  if (!appName) return res.sendStatus(400)
   try {
     const { path } = req.params
     if (!path) return res.sendStatus(400)
-    initFolder()
-    const fullPath = BASE_PATH_FILE + path.replaceAll('<', '/')
+    initFolder(appName)
+    const fullPath = `${PROJECT_PATH}/${appName}/js/${path}`
     if (!fs.existsSync(fullPath)) return res.sendStatus(404)
     fs.unlinkSync(fullPath)
     res.sendStatus(200)
